Here Fully Completed the Add property part
But there was an error that i need to figure out. while adding the rooms to database in the floors table there are extra entities creating automatically i need to figure it out sooner lets find it soon But i am mooving to other next phase that Vieving property Layout.